move wasbfs under datastore
